requirements.txt
django
python-social-auth
psycopg2
facebook-sdk